// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/observationStatus.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ObservationStatus = 0 | 1 | 2 | 3;

export const ObservationStatus = {
    NUMBER_0: 0 as ObservationStatus,
    NUMBER_1: 1 as ObservationStatus,
    NUMBER_2: 2 as ObservationStatus,
    NUMBER_3: 3 as ObservationStatus
};